Moved functions to new internal classes.
